# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lumentum Shared API
  description: |
    Merkezi API — Web, Android, iOS ve gelecekteki tüm platformlar bu sözleşmeyi kullanır.
    Tek beyin: `packages/api` + `packages/core-engine`.
  version: 0.1.0
  contact:
    name: Cicibyte
    url: https://lumentum.cicibyte.com
servers:
  - url: https://lumentum.cicibyte.com
    description: Production
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/reading/process:
    post:
      tags: [reading]
      summary: Metni RSVP token akışına dönüştür
      operationId: processReading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Token listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessReadingResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenData:
      $ref: "./schemas/token.json"
    ProcessReadingResponse:
      type: object
      required: [result, source]
      properties:
        result:
          type: array
          items:
            $ref: "#/components/schemas/TokenData"
        source:
          type: string
          enum: [rust, rust-cli, python-fallback]
...